creado la estrucutura de carpetas para mejor orden del proyectos
Aun falta arreglar la alerta cuando entramos a ventas, no lista los productos y clientes a menos que escribas en el buscador primero
