+++ tiendita/.htacces
@@ -1,0 +1,2 @@
+error_reporting(E_ALL);
+ini_set('display_errors', 1);
