--- tiendita/actualizarc.php
@@ -1,55 +1,0 @@
-<?php
-include('db.php');
-$con = conectar();
-
-$id = $_GET['id'];
-$sql = "select * from clientes where id='$id'";
-$query = mysqli_query($con, $sql);
-$row = mysqli_fetch_array($query);
-?>
-
-<!DOCTYPE html>
-<html lang="es">
-
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Actualizar</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM" crossorigin="anonymous"></script>
-    <link rel="stylesheet" href="productoss.css">
-    <style>
-        .w-80 {
-            width: 80%;
-        }
-    </style>
-</head>
-
-<body>
-    <center>
-        <div class="container w-80 bg-dark redondear position-absolute top-50 start-50 translate-middle">
-            <form action="updatec.php" method="post">
-                <div class="mt-3 text-start">
-                    <h3 class="text-center">Actualizar</h3>
-                    <input type="hidden" name="id" value="<?php echo $row['id'] ?>">
-                </div>
-                <div class="mt-4 text-start">
-                    <label class="for-label">Nombre</label>
-                    <input type="text" class="form-control" name="nombre" placeholder="ingrese nombre" value="<?php echo $row['nombre'] ?>">
-                </div>
-                <div class="mt-2 text-start mb-3">
-                    <div class="d-grid mt-4">
-                        <input type="submit" name="insertar" class="btn btn-success redondear mb-4" value="Actualizar Cliente">
-                    </div>
-            </form>
-        </div>
-    </center>
-    <script type="module" src="https://unpkg.com/ionicons@5.5.2/dist/ionicons/ionicons.esm.js"></script>
-    <script nomodule src="https://unpkg.com/ionicons@5.5.2/dist/ionicons/ionicons.js"></script>
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM" crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.9.2/dist/umd/popper.min.js" integrity="sha384-IQsoLXl5PILFhosVNubq5LC7Qb9DXgDA9i+tQ8Zj3iwWAwPtgFTxbJ8NT4GN1R8p" crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.min.js" integrity="sha384-cVKIPhGWiC2Al4u+LWgxfKTRIcfu0JTxR+EQDz/bgldoEyl4H0zUF0QKbrJ0EcQF" crossorigin="anonymous"></script>
-</body>
-
-</html>
