+++ tiendita/assets/js/code.js
@@ -1,0 +1,6 @@
+var hamburguer = document.getElementById('Hamburguer')
+var menu = document.getElementById('Menu')
+
+hamburguer.addEventListener("click", function(){
+    menu.classList.toggle('open');
+})
