+++ tiendita/backend/Crud-clientes/insertarc.php
@@ -1,0 +1,16 @@
+<?php
+include('../config/db.php');
+$con=conectar();
+
+
+    $nombre=$_POST['nombre'];
+    $sql="insert into clientes(nombre) values('$nombre')";
+    $query=mysqli_query($con,$sql);
+
+    if ($query) {
+        header("Location: clientes.php");
+
+    }else{
+
+    }
+?>
