+++ tiendita/backend/Crud-ventas/cuentas.php
@@ -1,0 +1,132 @@
+<!DOCTYPE html>
+<html lang="es">
+
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+    <!-- importante -->
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.3/jquery.min.js"></script>
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM" crossorigin="anonymous"></script>
+    <script src="//cdn.jsdelivr.net/npm/sweetalert2@11"></script>
+    <style>
+        body {
+            background-color: #000000;
+            color: #fff;
+        }
+
+        .producto {
+            max-width: 80px;
+        }
+
+        .fz-30 {
+            font-size: 30px;
+        }
+
+        .redondear {
+            border-radius: 16px;
+        }
+
+        #buscador {
+            position: sticky;
+            top: 0;
+            background-color: #000;
+        }
+    </style>
+
+</head>
+
+<body>
+    <center>
+        <?php include("../config/db.php");
+        $con = conectar(); ?>
+
+        <div id="buscador" class="container py-2">
+            <div class="d-flex mt-3 mb-2">
+                <div class="col">
+                    <input onkeyup="buscar_ahora($('#buscar_1').val());" type="text" class="form-control bg-dark text-light" id="buscar_1" name="buscar_1" placeholder="Buscar Producto">
+                </div>
+                <div class="col-auto ms-2">
+                    <a href="" class="btn btn-primary">
+                        <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-search" viewBox="0 0 16 16">
+                            <path d="M11.742 10.344a6.5 6.5 0 1 0-1.397 1.398h-.001c.03.04.062.078.098.115l3.85 3.85a1 1 0 0 0 1.415-1.414l-3.85-3.85a1.007 1.007 0 0 0-.115-.1zM12 6.5a5.5 5.5 0 1 1-11 0 5.5 5.5 0 0 1 11 0z" />
+                        </svg>
+                    </a>
+                </div>
+            </div>
+        </div>
+
+        <div class="container mt-2">
+
+            <div id="datos_buscador" class="row">
+
+            </div>
+
+        </div>
+    </center>
+
+
+
+    <script type="text/javascript">
+        function buscar_ahora(buscar) {
+            var parametros = {
+                "buscar": buscar
+            };
+            $.ajax({
+                data: parametros,
+                type: 'POST',
+                url: 'buscadorventas.php',
+                success: function(data) {
+                    document.getElementById("datos_buscador").innerHTML = data;
+                }
+            });
+        }
+        //   buscar_ahora();
+    </script>
+
+    <script>
+        function eliminacion(codigo) {
+            Swal.fire({
+                title: '¿Deseas eliminar?',
+                text: "¡No podra recuperarlo una vez eliminado!",
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: '¡Si, eliminar!',
+                cancelButtonText: 'Cancelar'
+            }).then((result) => {
+                if (result.isConfirmed) {
+                    mandar_php(codigo)
+                }
+            })
+        }
+    </script>
+    <script>
+        function mandar_php(codigo){
+            parametros = {id: codigo};
+            $.ajax({
+                data: parametros,
+                url: 'deletev.php',
+                type: 'POST',
+                beforeSend: function () {},
+                success: function () {
+                    Swal.fire('Eliminado!', 'La venta ah sido eliminada', 'success').then((result) =>{
+                        window.location.href = 'cuentas.php'
+                    });
+                },
+            });
+        }
+    </script>
+
+    <script type="module" src="https://unpkg.com/ionicons@5.5.2/dist/ionicons/ionicons.esm.js"></script>
+    <script nomodule src="https://unpkg.com/ionicons@5.5.2/dist/ionicons/ionicons.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM" crossorigin="anonymous"></script>
+    <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.9.2/dist/umd/popper.min.js" integrity="sha384-IQsoLXl5PILFhosVNubq5LC7Qb9DXgDA9i+tQ8Zj3iwWAwPtgFTxbJ8NT4GN1R8p" crossorigin="anonymous"></script>
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.min.js" integrity="sha384-cVKIPhGWiC2Al4u+LWgxfKTRIcfu0JTxR+EQDz/bgldoEyl4H0zUF0QKbrJ0EcQF" crossorigin="anonymous"></script>
+
+</body>
+
+</html>
