+++ tiendita/backend/config/config.php
@@ -1,0 +1,8 @@
+<?php
+
+$user = "root";
+$pass = "";
+$host = "localhost";
+$bd = "tiendita";
+
+?>
