--- tiendita/code.js
@@ -1,6 +1,0 @@
-var hamburguer = document.getElementById('Hamburguer')
-var menu = document.getElementById('Menu')
-
-hamburguer.addEventListener("click", function(){
-    menu.classList.toggle('open');
-})
