+++ tiendita/database/tienditabd.sql
