+++ tiendita/index.php
@@ -8,7 +8,7 @@
     <title>RIM</title>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM" crossorigin="anonymous"></script>
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="/assets/css/style.css">
 </head>
 
 <body>
@@ -17,7 +17,7 @@
         <div class="d-flex text-center mt-1">
             <h1 class="text-center">Inicio de sesion</h1>
         </div>
-        <form action="validar.php" method="POST">
+        <form action="backend/validaciones/validar.php" method="POST">
             <div class="mt-5 text-start">
                 <label class="for-label">Correo</label>
                 <input type="text" class="form-control" name="usuario" placeholder="usuario">
